feat (refs T25201): Extract duplicated methods in new class. And remove unused methods.
