+++ demosplan/DemosPlanCoreBundle/Logic/AssessmentTable/AssessmentTableServiceOutput.php
@@ -754,136 +754,6 @@ class AssessmentTableServiceOutput
     }
 
     /**
-     * Statement in unified data format.
-     *
-     * @return array - formatted statement
-     *
-     * @throws ReflectionException
-     *
-     * @deprecated Use {@link formatStatementObject} instead
-     */
-    public function formatStatementArray(array $statement): array
-    {
-        return [
-            'type'                          => 'statement',
-            'attachments'                   => $statement['attachments'] ?? null,
-            'authoredDate'                  => $statement['meta']['authoredDate'] ?? null,
-            'cluster'                       => $statement['cluster'] ?? null,
-            'documentTitle'                 => $statement['document']['title'] ?? null,
-            'externId'                      => $statement['externId'] ?? null,
-            'formerExternId'                => $statement['formerExternId'] ?? null,
-            'elementTitle'                  => $statement['element']['title'] ?? null,
-            'files'                         => $statement['files'] ?? null,
-            'orgaName'                      => $statement['meta']['orgaName'] ?? null,
-            'orgaDepartmentName'            => $statement['meta']['orgaDepartmentName'] ?? null,
-            'originalId'                    => $statement['original']['ident'] ?? null,
-            'paragraphTitle'                => $statement['paragraph']['title'] ?? null,
-            'parentId'                      => $statement['parent']['ident'] ?? null,
-            'polygon'                       => $statement['polygon'] ?? null,
-            'publicAllowed'                 => $statement['publicAllowed'] ?? null,
-            'publicCheck'                   => $statement['publicCheck'] ?? null,
-            'publicStatement'               => $statement['publicStatement'] ?? null,
-            'publicVerified'                => $statement['publicVerified'] ?? null,
-            'publicVerifiedTranslation'     => $statement['publicVerifiedTranslation'] ?? null,
-            'recommendation'                => $statement['recommendation'] ?? null,
-            'submit'                        => $statement['submit'] ?? null,
-            'submitName'                    => $statement['meta']['submitName'] ?? null,
-            'authorName'                    => $statement['meta']['authorName'] ?? null,
-            'text'                          => $statement['text'] ?? null,
-            'votes'                         => $statement['votes'] ?? null,
-            'votesNum'                      => $statement['votesNum'] ?? null,
-            'likesNum'                      => $statement['likesNum'] ?? null,
-            'fragments'                     => [],
-            'userState'                     => $statement['meta']['userState'] ?? null,
-            'userOrganisation'              => $statement['meta']['userOrganisation'] ?? null,
-            'userGroup'                     => $statement['meta']['userGroup'] ?? null,
-            'movedToProcedureName'          => $statement['movedToProcedureName'] ?? null,
-            'movedFromProcedureName'        => $statement['movedFromProcedureName'] ?? null,
-            'userPosition'                  => $statement['meta']['userPosition'] ?? null,
-            'isClusterStatement'            => $statement['isClusterStatement'] ?? null,
-            'name'                          => $statement['name'] ?? null,
-            'isSubmittedByCitizen'          => $statement['isSubmittedByCitizen'] ?? null,
-        ];
-    }
-
-    /**
-     * Statement in unified data format.
-     *
-     * @return array formatted statement
-     *
-     * @throws ReflectionException
-     */
-    public function formatStatementObject(Statement $statement): array
-    {
-        $item = $this->statementService->convertToLegacy($statement);
-        $item['parent'] = $this->statementService->convertToLegacy($statement->getParent());
-        $item['original'] = $this->statementService->convertToLegacy($statement->getOriginal());
-
-        return $this->formatStatementArray($item);
-    }
-
-    /**
-     * Fragment in unified data format.
-     *
-     * @return array - formatted fragment
-     *
-     * @throws Exception
-     */
-    public function formatFragmentObject(StatementFragment $fragment): array
-    {
-        $item = $this->formatStatementObject($fragment->getStatement());
-
-        // override selected item fields with fragment content:
-        $item['type'] = 'fragment';
-        $item['created'] = $fragment->getCreated(); // todo: check format
-        $item['recommendation'] = $fragment->getConsideration();
-        $item['text'] = $fragment->getText();
-        $item['elementId'] = $fragment->getElementId();
-        $item['elementTitle'] = $fragment->getElementTitle();
-
-        return $item;
-    }
-
-    /**
-     * Fragment in unified data format.
-     *
-     * @return array - formatted fragment
-     *
-     * @throws Exception
-     *
-     * @deprecated Use {@link formatFragmentObject} instead
-     */
-    public function formatFragmentArray(array $statement, array $fragment): array
-    {
-        $tmpElementId = $fragment['elementId'];
-        $tmpElementTitle = $fragment['elementTitle'];
-
-        $item = $this->formatStatementArray($statement);
-        $item['sortIndex'] = $fragment['sortIndex'];
-
-        // override selected item fields with fragment content:
-        $item['type'] = 'fragment';
-        $item['created'] = $fragment['created'] ?? null;
-
-        $item['text'] = '';
-        $item['recommendation'] = '';
-        // we need to fetch Fragment, as text fields are not mapped in statement
-        // index for performance reasons
-        $statementFragment = $this->statementHandler->getStatementFragment($fragment['id']);
-        if ($statementFragment instanceof StatementFragment) {
-            // pretend as if consideration would be an recommendation
-            // as it has the same behaviour
-            $item['recommendation'] = $statementFragment->getConsideration();
-            $item['text'] = $statementFragment->getText();
-        }
-
-        $item['elementId'] = $tmpElementId;
-        $item['elementTitle'] = $tmpElementTitle;
-
-        return $item;
-    }
-
-    /**
      * Collects Departments and usernames of clustered statements.
      *
      * @param array $item
